User service now stores data in a Map.
Previously the User service was depending on a nonexistent API to store
data. Instead of this, it now uses a Map to store aforementioned data.
In addition, a serialize method has been provided for persistance's sake. The
constructor has also been modified to allow for loading via an optional
parameter.